Eliminated warnings.

+++ debugger/external.c
@@ -117,7 +117,7 @@ void DEBUG_ExternalDebugger(void)
     memset(pid_string, 0, DBG_BUFF_SIZE);
 
     /* make pid into string */
-    sprintf(pid_string, "%d", attach_pid);
+    sprintf(pid_string, "%ld", (long) attach_pid);
 
     /* now exec the debugger to get it's own clean memory space */
     if (dbg_no_xterm)
